+++ src/app/main/map/components/map-menu/map-menu.component.ts
@@ -113,11 +113,11 @@ export class MapMenuComponent implements OnInit, OnChanges {
     );
 
     this.reportsNotClosed = this.reportsMonthly.filter(
-      report => report.status === ReportSatus.CUT
+      report => report.status === ReportSatus.ONGOING
     );
 
     this.reportsDayNotClosed = this.reportsDay.filter(
-      report => report.status === ReportSatus.CUT
+      report => report.status === ReportSatus.ONGOING
     );
   }
 
@@ -128,8 +128,8 @@ export class MapMenuComponent implements OnInit, OnChanges {
     // the sort by update date
     this.reportsMonthly.sort(
       (a: any, b: any) => {
-        const aDate = a?._updatedAt?.seconds;
-        const bDate = b?._updatedAt?.seconds;
+        const aDate = a?.updatedAt?.seconds;
+        const bDate = b?.updatedAt?.seconds;
 
         if (aDate < bDate) {
           return 1;