+++ src/environments/environment.ts
@@ -2,7 +2,7 @@ export const environment = {
   production: false,
   environment: 'dev',
   domain: 'localhost',
-  stadiaApiKey: '',
+  stadiaApiKey: '97565f4c-6ad7-4b06-b36c-eefe81b1fed1',
   app: {
   modules: {
   mapMenu: true,
@@ -12,14 +12,14 @@ export const environment = {
   }
   },
   firebase: {
-    apiKey: 'FIREBASE_API_KEY',
-    authDomain: 'FIREBASE_AUTH_DOMAIN',
-    databaseURL: 'FIREBASE_DATABASE_URL',
-    messagingSenderId: 'FIREBASE_MESSAGING_SENDER_ID',
-    appId: 'FIREBASE_APP_ID',
-    measurementId: 'FIREBASE_MEASUREMENT_ID',
-    projectId: 'FIREBASE_PROJECT_ID',
-    storageBucket: 'FIREBASE_STORAGE_BUCKET'
+    apiKey: 'AIzaSyDgxARuTPtogkJCFyCOK3kcgkfXmdgbkgA',
+    authDomain: 'lightcutoff-dev.firebaseapp.com',
+    databaseURL: 'https://lightcutoff-dev.firebaseio.com',
+    messagingSenderId: '708077956449',
+    appId: '1:708077956449:web:8307c1ccd10dd895726920',
+    measurementId: 'G-5MMMZRR7G0',
+    projectId: 'lightcutoff-dev',
+    storageBucket: 'lightcutoff-dev.appspot.com'
   }
   };